back to mutable tree
